Fitting refactor
